# Fix performance of account timelines