When `exclude_reblogs`, `tagged` or `only_media` filters are requested, there is no way that a reblog would turn up in the results; nevertheless, previously a join was performed even on those queries to filter out potential reblogs from excluded accounts. This made those queries unnecessarily slower.

Comparison for a "Posts" tab query as a logged-in follower (same profile):

|Server|Before|After|
|-|-|-|
|mastodon.online (93 GB)|32.9ms|3.9ms|
|mastodon.social (401 GB)|829.9ms|6.5ms|

Comparison for the "Media" tab (same profile):

|Server|Before|After|
|-|-|-|
|mastodon.online (93 GB)|128ms|4ms|
|mastodon.social (401 GB)|18s|11ms|