@Gargron 2022-03-07
No more behaviour change regarding mentions, join is faster than subquery. Fixed typo.